Web_Screping-3
Sir, I'm able to scrab only a list of page_content, some information about (tutorial, reference, example),& links. Now what to do, not able to get the table & rows & headings, kindly help with this. i tried to create a data table but was not able to find the proper solution.
kindly Check once if there is an error in my code you find please help to resolve the error.

Thank You..